발주완료 화면에서 ‘홈으로 이동’ 버튼 클릭하여 홈 화면으로 이동 후 로그아웃하고 다른 계정으로 재로그인하려고 하면 로그인이 안되는 이슈를 해결하는 작업 완성 / 마이페이지 화면 기초 프레임 재설계 작업 완성 - 기존의 파이어베이스 데이터 업데이트 하는 로직은 드로워화면으로 이주 (cart_repository.dart 내 getCartItemCount() 함수, uploadImage 함수, addToCartItem 함수, getCartItems() 함수, updateCartItemQuantity 함수, removeCartItem 함수, cartItemsStream() 함수, updateCartItemChecked 함수에 로그인 상태 확인 및 예외 처리 관련 로직 추가 / cart_state_provider.dart 내 _subscribeToCartItems() 함수, loadCartItems() 함수에 사용자 인증 체크를 하는 로직 추가 - 즉, 해당 장바구니 관련 레퍼지토리 및 상태관리 함수들이 로그인 상태를 확인하지 않다보니 문제 발생했고, 이를 개선!!!!) (banner_model.dart 내 배너 이미지의 URL을 저장하기 위한 모델 클래스인 ProfileMainSmall1BannerImage 생성 / banner_repository.dart 내 Firestore 데이터베이스에서 배너 이미지 데이터를 가져오는 기능인 ProfileMainSmall1BannerRepository 생성 / product_all_providers.dart 내 Firestore에서 작은 배너1 이미지 정보를 가져오기 위한 레포지토리 클래스의 인스턴스를 생성하는 프로바이더인 profileMainSmall1BannerRepositoryProvider와 앱에서 사용되는 여러 배너 이미지들을 Firestore로부터 받아와서 리스트 형태로 제공하는 profileMainSmall1BannerImagesProvider 생성 / profile_state_provider.dart 내 profileMainSmall1BannerPageProvider와 profileMainScrollPositionProvider인 화면 자체의 스크롤 상태관리와 작은 소배너 상태관리 함수 생성 / product_body_parts_layout.dart 내 logoutAndLoginAfterProviderReset에서 마이페이지 화면 관련 profileMainScrollPositionProvider와 profileMainSmall1BannerPageProvider 상태관리 provider 초기화 로직 추가 / profile_screen.dart 내 작은 소배너1이 있는 마이페이지 화면 기초 프레임 관련 로직 구현)) (추후, 마이페이지 화면 세부 설계 작업 시작 필요!!)(240801 [22:43])

+++ lib/cart/repository/cart_repository.dart
@@ -4,6 +4,7 @@ import 'package:cloud_firestore/cloud_firestore.dart'; // Firebase Firestore를 
 import 'package:firebase_storage/firebase_storage.dart'; // Firebase Storage를 사용하기 위해 import
 import '../../product/model/product_model.dart'; // 제품 모델을 사용하기 위해 import
 
+
 // ------- 장바구니와 관련된 데이터를 Firebase에 저장하고 저장된 데이터를 불러오고 하는 관리 관련 데이터 처리 로직인 CartItemRepository 클래스 시작
 // CartItemRepository 클래스는 Firestore와의 데이터 통신을 담당하는 역할
 class CartItemRepository {
@@ -15,6 +16,10 @@ class CartItemRepository {
 
   // Firestore에서 장바구니 아이템 개수 가져오는 함수 - Firestore에서 'cart_item' 컬렉션의 문서 개수를 반환
   Future<int> getCartItemCount() async {
+    final user = FirebaseAuth.instance.currentUser; // 현재 로그인한 사용자 정보 가져옴
+    if (user == null) { // 사용자가 로그인되어 있지 않은 경우 예외 발생
+      throw Exception('User not logged in'); // 예외 발생
+    }
     final userId = FirebaseAuth.instance.currentUser!.uid; // 현재 로그인한 사용자 ID 가져옴
     final querySnapshot = await firestore.collection('cart_item').doc(userId).collection('items').get(); // 'cart_item' 컬렉션의 모든 문서를 가져옴
     return querySnapshot.docs.length; // 문서의 개수를 반환
@@ -22,7 +27,11 @@ class CartItemRepository {
 
   // 이미지 URL을 Firebase Storage에 업로드하는 함수 - 주어진 이미지 URL을 가져와 Firebase Storage에 저장하고, 다운로드 URL을 반환
   Future<String> uploadImage(String imageUrl, String storagePath) async {
-    final userId = FirebaseAuth.instance.currentUser!.uid; // 현재 로그인한 사용자 ID 가져옴
+    final user = FirebaseAuth.instance.currentUser; // 현재 로그인한 사용자 정보 가져옴
+    if (user == null) { // 사용자가 로그인되어 있지 않은 경우 예외 발생
+      throw Exception('User not logged in'); // 예외 발생
+    }
+    final userId = user.uid; // 현재 로그인한 사용자 ID 가져옴
     final response = await http.get(Uri.parse(imageUrl)); // 주어진 이미지 URL로부터 데이터를 가져옴
     final bytes = response.bodyBytes; // 이미지 데이터를 바이트로 변환
     final ref = storage.ref().child('$userId/$storagePath'); // Firebase Storage에 저장할 경로 생성
@@ -34,7 +43,11 @@ class CartItemRepository {
   // 장바구니에 아이템 추가하는 함수 - 선택된 색상, 사이즈, 수량 정보를 포함하여 장바구니 아이템을 Firestore에 추가
   Future<void> addToCartItem(ProductContent product, String? selectedColorText,
       String? selectedColorUrl, String? selectedSize, int quantity) async {
-    final userId = FirebaseAuth.instance.currentUser!.uid; // 현재 로그인한 사용자 ID 가져옴
+    final user = FirebaseAuth.instance.currentUser; // 현재 로그인한 사용자 정보 가져옴
+    if (user == null) { // 사용자가 로그인되어 있지 않은 경우 예외 발생
+      throw Exception('User not logged in'); // 예외 발생
+    }
+    final userId = user.uid; // 현재 로그인한 사용자 ID 가져옴
 
     // Firestore에 저장할 데이터 준비
     final data = {
@@ -78,7 +91,11 @@ class CartItemRepository {
   Future<List<Map<String, dynamic>>> getCartItems() async {
     // 'cart_item' 컬렉션에서 모든 문서를 'timestamp' 필드 기준으로 내림차순 정렬하여 가져옴
     // - 장바구니 화면 내 timestamp인 필드 데이터를 가지고 내림차순으로 상품 데이터를 정렬하여 최신 데이터가 제일 상단에 위치하도록 함
-    final userId = FirebaseAuth.instance.currentUser!.uid; // 현재 로그인한 사용자 ID 가져옴
+    final user = FirebaseAuth.instance.currentUser; // 현재 로그인한 사용자 정보 가져옴
+    if (user == null) { // 사용자가 로그인되어 있지 않은 경우 예외 발생
+      throw Exception('User not logged in'); // 예외 발생
+    }
+    final userId = user.uid; // 현재 로그인한 사용자 ID 가져옴
     final querySnapshot = await firestore.collection('cart_item').doc(userId).collection('items').orderBy('timestamp', descending: true).get(); // 'cart_item' 컬렉션에서 모든 문서를 'timestamp' 필드 기준으로 내림차순 정렬하여 가져옴
     return querySnapshot.docs.map((doc) { // 가져온 문서들을 순회하여 리스트로 반환
       final data = doc.data(); // 문서 데이터를 가져옴
@@ -89,21 +106,33 @@ class CartItemRepository {
 
   // Firestore에서 가져온 수량 데이터를 업데이트하는 함수
   Future<void> updateCartItemQuantity(String docId, int newQuantity) async {
-    final userId = FirebaseAuth.instance.currentUser!.uid; // 현재 로그인한 사용자 ID 가져옴
+    final user = FirebaseAuth.instance.currentUser; // 현재 로그인한 사용자 정보 가져옴
+    if (user == null) { // 사용자가 로그인되어 있지 않은 경우 예외 발생
+      throw Exception('User not logged in'); // 예외 발생
+    }
+    final userId = user.uid; // 현재 로그인한 사용자 ID 가져옴
     await firestore.collection('cart_item').doc(userId).collection('items').doc(docId).update({ // 주어진 문서 ID에 해당하는 문서의 'selected_count' 필드를 업데이트함
-      'selected_count': newQuantity,
+      'selected_count': newQuantity, // 새로운 수량으로 업데이트
     });
   }
 
   // 장바구니 화면 내에서 상품 아이템을 '삭제' 버튼 클릭 시, Firestore에서 삭제되도록 하는 함수
   Future<void> removeCartItem(String docId) async {
-    final userId = FirebaseAuth.instance.currentUser!.uid; // 현재 로그인한 사용자 ID 가져옴
+    final user = FirebaseAuth.instance.currentUser; // 현재 로그인한 사용자 정보 가져옴
+    if (user == null) { // 사용자가 로그인되어 있지 않은 경우 예외 발생
+      throw Exception('User not logged in'); // 예외 발생
+    }
+    final userId = user.uid; // 현재 로그인한 사용자 ID 가져옴
     await firestore.collection('cart_item').doc(userId).collection('items').doc(docId).delete(); // 주어진 문서 ID에 해당하는 문서를 Firestore에서 삭제
   }
 
   // Firestore에서 장바구니 아이템 스트림을 가져오는 함수
   Stream<List<Map<String, dynamic>>> cartItemsStream() {
-    final userId = FirebaseAuth.instance.currentUser!.uid; // 현재 로그인한 사용자 ID 가져옴
+    final user = FirebaseAuth.instance.currentUser; // 현재 로그인한 사용자 정보 가져옴
+    if (user == null) { // 사용자가 로그인되어 있지 않은 경우 예외 발생
+      throw Exception('User not logged in'); // 예외 발생
+    }
+    final userId = user.uid; // 현재 로그인한 사용자 ID 가져옴
     // Firestore의 'cart_item' 컬렉션을 timestamp 내림차순으로 정렬하여 가져옴.
     return firestore.collection('cart_item').doc(userId).collection('items').orderBy('timestamp', descending: true).snapshots().map((querySnapshot) {
       // 쿼리 결과를 문서 목록으로 변환.
@@ -121,8 +150,12 @@ class CartItemRepository {
 
   // Firestore에서 특정 아이템의 체크 상태를 업데이트하는 함수인 updateCartItemChecked
   Future<void> updateCartItemChecked(String id, bool checked) async {
-    final userId = FirebaseAuth.instance.currentUser!.uid; // 현재 로그인한 사용자 ID 가져옴
-    await firestore.collection('cart_item').doc(userId).collection('items').doc(id).update({'bool_checked': checked});
+    final user = FirebaseAuth.instance.currentUser; // 현재 로그인한 사용자 정보 가져옴
+    if (user == null) { // 사용자가 로그인되어 있지 않은 경우 예외 발생
+      throw Exception('User not logged in'); // 예외 발생
+    }
+    final userId = user.uid; // 현재 로그인한 사용자 ID 가져옴
+    await firestore.collection('cart_item').doc(userId).collection('items').doc(id).update({'bool_checked': checked}); // 주어진 ID에 해당하는 문서의 체크 상태를 업데이트
   }
 }
 // ------- 장바구니와 관련된 데이터를 Firebase에 저장하고 저장된 데이터를 불러오고 하는 관리 관련 데이터 처리 로직인 CartItemRepository 클래스 끝
